+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changes
 
 Unreleased
 ----------
+- Changed how `now` is imported from `timezone` (`timezone` module is imported now)
 - Add simple filtering if provided a minutes argument in `clean_duplicate_history` (gh-606)
 - Add setting to convert `FileField` to `CharField` instead of `TextField` (gh-623)
 - import model `ContentType` in `SimpleHistoryAdmin` using `django_apps.get_model`
@@ -10,6 +11,8 @@ Unreleased
 - Fix `utils.update_change_reason` when user specifies excluded_fields
 - Render fields as readonly in history detail view if `SIMPLE_HISTORY_EDIT` is not set
   `True`
+- settings.SIMPLE_HISTORY_REVERT_DISABLED if True removes the Revert
+  button from the history form for all historical models (gh-632))
 
 2.8.0 (2019-12-02)
 ------------------